requirements.txt
requests-aws4auth==1.1.1
boto3==1.20.31
elasticsearch==7.13.4
python-dotenv==0.19.2
